fix: rename common files
