+++ src/common/enums/magic-signs.ts
