+++ src/common/interfaces/card.ts
