+++ src/common/interfaces/index.ts
@@ -1,1 +1,1 @@
-export { ICard } from './ICard';
+export { ICard } from './card';
